Create separate project component and adjust for mobile view
